use integer math to get the fraction. doubles are emulated and supposedly floats are not, but if I use a float and look at the assembly it's still calling into the emulation library, so just don't bother.

+++ main/esp-ntp.cpp
@@ -303,7 +303,7 @@ static bool stamp(ntp_timestamp *ts) {
         vPortExitCritical();
         return false;
     }
-    ts->fraction = uint32_t(double(usec) * (1LL << 32) / 1'000'000);
+    ts->fraction = uint32_t((uint64_t(usec) * (1LL << 32)) / 1'000'000);
     ts->seconds = htonl(ts->seconds);
     ts->fraction = htonl(ts->fraction);
     return true;
